refactor: group LoginController dependencies an a single file
